==> notebooks/05_fusion_enrichissement.ipynb <==
from pyspark.sql import SparkSession
import pandas as pd
from pyspark.sql import functions as F
from pathlib import Path

DATA_DIR = Path("../data_ecf").resolve()
OUTPUT_DIR = Path("../output").resolve()
CONSOMMATIONS_CLEAN_PATH = OUTPUT_DIR / "consommation_clean"
BATIMENTS_PATH = DATA_DIR / "batiments.csv"
CONSOMMATIONS_AGREGEES_PATH = OUTPUT_DIR / "consommations_agregees"
METEO_CLEAN_PATH = OUTPUT_DIR / "meteo_clean.csv"
TARIFS_ENERGIE_PATH = DATA_DIR / "tarifs_energie.csv"
CONSOMMATIONS_ENRICHIES_CSV_PATH = OUTPUT_DIR / "consommations_enrichies.csv"
CONSOMMATIONS_ENRICHIES_PARQUET_PATH = OUTPUT_DIR / "consommations_enrichies"

#### Etape 2.2 : Fusion et enrichissement
- Charger les consommations nettoyees (depuis Parquet)

df_consommations = pd.read_parquet(CONSOMMATIONS_AGREGEES_PATH.as_posix())
df_consommations['timestamp'] = pd.to_datetime(df_consommations['timestamp'])
df_consommations['date'] = pd.to_datetime(df_consommations['date'])
df_consommations['type_energie'] = df_consommations['type_energie'].astype("str")

df_consommations.info()
df_consommations.head(10)

- Fusionner avec les donnees meteo (sur commune et timestamp arrondi a l'heure)

df_meteo = pd.read_csv(METEO_CLEAN_PATH.as_posix())
df_meteo['timestamp'] = pd.to_datetime(df_meteo['timestamp'])
df_meteo['date'] = pd.to_datetime(df_meteo['date'])

df_meteo.info()
df_meteo.head(10)

df_consommations_meteo = df_consommations.merge(
        df_meteo,
        on=["commune", "timestamp", "date", "month"],
        how="left",
)

print(f"  Nombre de colonnes après fusion : {len(df_consommations_meteo.columns)}")
print(df_consommations_meteo.columns)

- Fusionner avec le referentiel batiments

La fusion avec le référentiel batiements a déjà été réalisée dans le fichier `03_agregations_spark.py`.

- Fusionner avec les tarifs pour calculer le cout financier

df_tarifs_energie = pd.read_csv(TARIFS_ENERGIE_PATH.as_posix())
df_tarifs_energie['date_debut'] = pd.to_datetime(df_tarifs_energie['date_debut'])
df_tarifs_energie['date_fin'] = pd.to_datetime(df_tarifs_energie['date_fin'])
df_tarifs_energie['type_energie'] = df_tarifs_energie['type_energie'].astype("str")

df_tarifs_energie.info()
df_tarifs_energie.head(10)

df_merge = (
    df_consommations_meteo
    .merge(df_tarifs_energie, on="type_energie", how="left")
    .query("date >= date_debut and date <= date_fin")
)

print(f"  Nombre de colonnes après fusion : {len(df_merge.columns)}")
print(df_merge.columns)
print(f"  Nombre de lignes avant jointure : {len(df_consommations_meteo)}")
print(f"  Nombre de lignes après jointure : {len(df_merge)}")

- Creer des features derivees :
  - Consommation par occupant
  - Consommation par m2
  - Cout journalier, mensuel, annuel
  - Indice de performance energetique (IPE)
  - Ecart a la moyenne de la categorie

df_features = df_merge.copy()

# Consommation par occupant
df_features["consommation_par_occupant"] = df_features["consommation"] / df_features["nb_occupants_moyen"]

# Consommation par m2
df_features["consommation_par_m2"] = df_features["consommation"] / df_features["surface_m2"]

# Cout journalier, mensuel et annuel
consommation_journaliere = df_features.groupby(["batiment_id", "date", "type_energie", "tarif_unitaire"]).agg(consommation_journaliere=("consommation", "mean")).reset_index()
consommation_journaliere["cout_journalier"] = consommation_journaliere["consommation_journaliere"] * consommation_journaliere["tarif_unitaire"]
df_features = df_features.merge(consommation_journaliere, on=["batiment_id", "date", "type_energie", "tarif_unitaire"], how="left")

cout_mensuel = df_features.groupby(["batiment_id", "year", "month", "type_energie"]).agg(cout_mensuel=("cout_journalier", "sum"))
cout_annuel = df_features.groupby(["batiment_id", "year", "type_energie"]).agg(cout_annuel=("cout_journalier", "sum"))
df_features = (
    df_features
    .join(cout_mensuel, ["batiment_id", "year", "month", "type_energie"], "left")
    .join(cout_annuel, ["batiment_id", "year", "type_energie"], "left")
)

# Indice de performance énergétique
consommation_annuelle = df_features.groupby(["batiment_id", "surface_m2", "year", "type_energie"]).agg(consommation_annuelle=("consommation_journaliere", "sum")).reset_index()
consommation_annuelle["IPE"] = consommation_annuelle["consommation_annuelle"] / consommation_annuelle["surface_m2"]
df_features = df_features.merge(consommation_annuelle, on=["batiment_id", "surface_m2", "year", "type_energie"], how="left")

# Ecart à la moyenne des catégorie
consommation_moyenne_par_type = df_features.groupby(["type", "type_energie"]).agg(consommation_moyenne_par_type=("consommation", "mean"))
df_features = df_features.join(consommation_moyenne_par_type, ["type", "type_energie"], "left")
df_features["ecart_conso_moyenne_type"] = df_features["consommation"] / df_features["consommation_moyenne_par_type"]

print(f"  Nombre de colonnes après ajout des features : {len(df_merge.columns)}")
print(df_features.columns)

J'ai concidéré que la consommation journalière était la médiane des consommation horraire. Ce qui n'est qu'une approximation.

# Sauvegarde CSV
df_features.to_csv(CONSOMMATIONS_ENRICHIES_CSV_PATH.as_posix(), index=False)

# Sauvegarde PARQUET
df_features.to_parquet(CONSOMMATIONS_ENRICHIES_PARQUET_PATH.as_posix(), index=False)

- Dictionnaire de donnees (description de toutes les colonnes)

Colonne | Description
:- | :-
'batiment_id' | Identifiant des bâtiments (BAT00056)
'timestamp' | Timestamp de la mesure de consommation (2024-01-13 08:00:00)
'consommation' | Consommation énergétique
'unite' | Unité de la mesure ('m3', 'kWh')
'hour' | Heure de la mesure
'year' | Année de la mesure
'month' | Mois de la mesure
'date' | Date de la mesure, format 2024-01-13
'nom' | Nom du bâtiment
'type' | Type de bâtiment ('ecole', 'mairie', 'mediatheque', 'gymnase', 'piscine')
'commune' | Nom de la commune
'surface_m2' | suface en m2 du bâtiment
'annee_construction' | année de construction du bâtiment
'classe_energetique' | Classe énergétique du bâtiment
'nb_occupants_moyen' | Nombre d'occupants moyen du bâtiment
'intensite_energetique' | Quotient consommation / surface_m2
'type_energie' | Type d'énergie ('eau', 'electricite', 'gaz')
'temperature_c' | Température météo correspondant au timestamp
'humidite_pct' | Humidité correspondante au timestamp
'rayonnement_solaire_wm2' | Rayonnement correspondant au timestamp
'vitesse_vent_kmh' | Vitesse du vent correspondant au timestamp
'precipitation_mm' | Précipitation correspondant au timestamp
'day_of_week' | Jour le de semaine de la mesure
'season' | Saison de la mesure
'date_debut' | Date de début du tarif unitaire
'date_fin' | Date de fin du tarif unitaire
'tarif_unitaire' | Tarif unitaire
'consommation_par_occupant' | Quotient consommation / nb_occupants_moyen
'consommation_par_m2' | Quotient consommation / surface_m2
'tarif_journalier' | Produit consommation_journalière * tarif_unitaire
'consommation_journaliere' | Consommation moyenne journalière
'cout_journialier' | Coût journalier d'un bâtiment
'cout_mensuel' | Coût mentuel d'un bâtiment
'cout_annuel' | Coût annuel d'un bâtiment
'consommation_annuelle' | Consommation annuelle d'un bâtiment
'IPE' | Indice de performance énergétique, quotient consommation_annuelle / surface_m2
'consommation_moyenne_par_type' | Consommation annuelle par type de bâtiment
'ecart_conso_moyenne_type' | Ecart à la consommation annuelle par type de bâtiment
